chore: java class to response/request body

+++ src/main/java/com/prog3td/demo/model/Match.java
@@ -28,7 +28,7 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 public class Match {
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    private int idPlay;
+    private Long idPlay;
     private LocalDateTime dateTime;
     private String stadium;
     @OneToOne
